requirements.txt
# ⚙️ Backend Web Framework
fastapi==0.115.9
uvicorn==0.21.1
python-multipart==0.0.9
pydantic==2.10.4

# 🔒 Authentication
bcrypt==4.3.0

# 📄 File Parsing & OCR
PyMuPDF==1.23.21
PyPDF2==3.0.1
pdfminer.six==20221105
pytesseract==0.3.10
pillow==11.0.0
opencv-python==4.9.0.80
pdf2image==1.17.0
numpy==2.2.6

# 📊 Data Handling
pandas>=1.3.0,<3.0
python-dotenv==1.0.0

# 📦 Embedding / LLM Integration
sentence-transformers==4.1.0
torch==2.7.0
langchain==0.2.2
langchain-openai==0.1.7
langchain-community==0.0.38
openai==1.24.0
replicate==0.24.0
chromadb==1.0.10

# 📦 Misc
tqdm==4.67.1
